pyproject.toml
[project]
name = "ai-phone-caller"
version = "0.1.0"
description = ""
authors = [
    {name = "Nehil Jain", email = "2761999+nehiljain@users.noreply.github.com"},
]
dependencies = [
    "modal-client>=0.51.3192",
    "jupyter>=1.0.0",
    "numpy>=1.25.2",
    "pandas>=2.0.3",
    "matplotlib>=3.7.2",
    "scipy>=1.9.3",
    "seaborn>=0.12.2",
    "cohere>=4.21",
    "langchain>=0.0.273",
    "openai>=0.27.9",
    "pymongo>=4.5.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
